pyproject.toml
[tool.poetry]
name = "sensor"
version = "0.1.0"
description = ""
authors = ["seongcheol"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
adafruit-dht = "^1.4.0"
pymysql = "^1.1.0"
apscheduler = "^3.10.4"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
